+++ hbservice-core/src/main/java/com/cmic/hbservice/service/OrderService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface OrderService {
     void addOrder(Order order);
     List<Order> getPendingList();
-    void cancelOrder(int orderId);
+    void cancelOrder(long orderId);
 }
